refactor: extract swap tokens to alloyed asset to mod
